Move Mockito-based mocks to own file
