+++ tests/src/test/scala/akka/kafka/internal/ConsumerTest.scala
@@ -5,8 +5,7 @@
 
 package akka.kafka.internal
 
-import java.util.concurrent.TimeUnit
-
+import akka.Done
 import akka.actor.ActorSystem
 import akka.kafka.ConsumerMessage._
 import akka.kafka.scaladsl.Consumer
@@ -17,17 +16,11 @@ import akka.stream.scaladsl._
 import akka.stream.testkit.scaladsl.StreamTestKit.assertAllStagesStopped
 import akka.stream.testkit.scaladsl.TestSink
 import akka.testkit.TestKit
-import akka.{Done, NotUsed}
 import org.apache.kafka.clients.consumer._
 import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.errors.WakeupException
 import org.apache.kafka.common.serialization.StringDeserializer
-import org.mockito
-import org.mockito.Mockito
 import org.mockito.Mockito._
-import org.mockito.invocation.InvocationOnMock
-import org.mockito.stubbing.Answer
-import org.mockito.verification.VerificationMode
 import org.scalatest.{BeforeAndAfterAll, FlatSpecLike, Matchers}
 
 import scala.collection.JavaConverters._
@@ -118,13 +111,6 @@ class ConsumerTest(_system: ActorSystem)
                                groupId: String = "group1",
                                topics: Set[String] = Set("topic")): Source[CommittableMessage[K, V], Control] =
     Consumer.commitWithMetadataSource(testSettings(mock, groupId), Subscriptions.topics(topics), metadataFromRecord)
-
-  def createCommittablePartitionedSource(
-      mock: Consumer[K, V],
-      groupId: String = "group1",
-      topics: Set[String] = Set("topic")
-  ): Source[(TopicPartition, Source[CommittableMessage[K, V], NotUsed]), Control] =
-    Consumer.committablePartitionedSource(testSettings(mock, groupId), Subscriptions.topics(topics))
 
   it should "fail stream when poll() fails with unhandled exception" in {
     assertAllStagesStopped {
@@ -739,141 +725,4 @@ class ConsumerTest(_system: ActorSystem)
       mock.verifyClosed()
     }
   }
-}
-
-object ConsumerMock {
-  type CommitHandler = (Map[TopicPartition, OffsetAndMetadata], OffsetCommitCallback) => Unit
-
-  def notImplementedHandler: CommitHandler = (_, _) => ???
-
-  class LogHandler extends CommitHandler {
-    var calls: Seq[(Map[TopicPartition, OffsetAndMetadata], OffsetCommitCallback)] = Seq.empty
-    def apply(offsets: Map[TopicPartition, OffsetAndMetadata], callback: OffsetCommitCallback) =
-      calls :+= ((offsets, callback))
-  }
-}
-
-class ConsumerMock[K, V](handler: ConsumerMock.CommitHandler = ConsumerMock.notImplementedHandler) {
-  private var responses = collection.immutable.Queue.empty[Seq[ConsumerRecord[K, V]]]
-  private var pendingSubscriptions = List.empty[(List[String], ConsumerRebalanceListener)]
-  private var assignment = Set.empty[TopicPartition]
-  private var messagesRequested = false
-  val mock = {
-    val result = Mockito.mock(classOf[KafkaConsumer[K, V]])
-    Mockito
-      .when(result.poll(mockito.ArgumentMatchers.any[Long]))
-      .thenAnswer(new Answer[ConsumerRecords[K, V]] {
-        override def answer(invocation: InvocationOnMock) = ConsumerMock.this.synchronized {
-          pendingSubscriptions.foreach {
-            case (topics, callback) =>
-              val tps = topics.map { t =>
-                new TopicPartition(t, 1)
-              }
-              assignment ++= tps
-              callback.onPartitionsAssigned(tps.asJavaCollection)
-          }
-          pendingSubscriptions = List.empty
-          val records = if (messagesRequested) {
-            responses.dequeueOption
-              .map {
-                case (element, remains) =>
-                  responses = remains
-                  element
-                    .groupBy(x => new TopicPartition(x.topic(), x.partition()))
-                    .map {
-                      case (topicPart, messages) => (topicPart, messages.asJava)
-                    }
-              }
-              .getOrElse(Map.empty)
-          } else Map.empty[TopicPartition, java.util.List[ConsumerRecord[K, V]]]
-          new ConsumerRecords[K, V](records.asJava)
-        }
-      })
-    Mockito
-      .when(
-        result.commitAsync(mockito.ArgumentMatchers.any[java.util.Map[TopicPartition, OffsetAndMetadata]],
-                           mockito.ArgumentMatchers.any[OffsetCommitCallback])
-      )
-      .thenAnswer(new Answer[Unit] {
-        override def answer(invocation: InvocationOnMock) = {
-          val offsets = invocation.getArgument[java.util.Map[TopicPartition, OffsetAndMetadata]](0)
-          val callback = invocation.getArgument[OffsetCommitCallback](1)
-          handler(offsets.asScala.toMap, callback)
-          ()
-        }
-      })
-    Mockito
-      .when(
-        result.subscribe(mockito.ArgumentMatchers.any[java.util.List[String]],
-                         mockito.ArgumentMatchers.any[ConsumerRebalanceListener])
-      )
-      .thenAnswer(new Answer[Unit] {
-        override def answer(invocation: InvocationOnMock) = {
-          val topics = invocation.getArgument[java.util.List[String]](0)
-          val callback = invocation.getArgument[ConsumerRebalanceListener](1)
-          pendingSubscriptions :+= (topics.asScala.toList -> callback)
-          ()
-        }
-      })
-    Mockito
-      .when(result.resume(mockito.ArgumentMatchers.any[java.util.Collection[TopicPartition]]))
-      .thenAnswer(new Answer[Unit] {
-        override def answer(invocation: InvocationOnMock) = {
-          messagesRequested = true
-          ()
-        }
-      })
-    Mockito
-      .when(result.pause(mockito.ArgumentMatchers.any[java.util.Collection[TopicPartition]]))
-      .thenAnswer(new Answer[Unit] {
-        override def answer(invocation: InvocationOnMock) = {
-          messagesRequested = false
-          ()
-        }
-      })
-    Mockito
-      .when(result.assignment())
-      .thenAnswer(new Answer[java.util.Set[TopicPartition]] {
-        override def answer(invocation: InvocationOnMock) = assignment.asJava
-      })
-    result
-  }
-
-  def enqueue(records: Seq[ConsumerRecord[K, V]]) =
-    synchronized {
-      responses :+= records
-    }
-
-  def verifyClosed(mode: VerificationMode = Mockito.times(1)) =
-    verify(mock, mode).close(ConsumerTest.closeTimeout.toMillis, TimeUnit.MILLISECONDS)
-
-  def verifyPoll(mode: VerificationMode = Mockito.atLeastOnce()) =
-    verify(mock, mode).poll(mockito.ArgumentMatchers.any[Long])
-
-  def assignPartitions(tps: Set[TopicPartition]) =
-    tps.groupBy(_.topic()).foreach {
-      case (topic, localTps) =>
-        pendingSubscriptions.find(_._1 == topic).get._2.onPartitionsAssigned(localTps.asJavaCollection)
-    }
-
-  def revokePartitions(tps: Set[TopicPartition]) =
-    tps.groupBy(_.topic()).foreach {
-      case (topic, localTps) =>
-        pendingSubscriptions.find(_._1 == topic).get._2.onPartitionsRevoked(localTps.asJavaCollection)
-    }
-}
-
-class FailingConsumerMock[K, V](throwable: Throwable, failOnCallNumber: Int*) extends ConsumerMock[K, V] {
-  var callNumber = 0
-
-  Mockito
-    .when(mock.poll(mockito.ArgumentMatchers.any[Long]))
-    .thenAnswer(new Answer[ConsumerRecords[K, V]] {
-      override def answer(invocation: InvocationOnMock) = FailingConsumerMock.this.synchronized {
-        callNumber = callNumber + 1
-        if (failOnCallNumber.contains(callNumber))
-          throw throwable
-        else new ConsumerRecords[K, V](Map.empty[TopicPartition, java.util.List[ConsumerRecord[K, V]]].asJava)
-      }
-    })
 }
